requirements.txt
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
aiofiles>=23.2.1
pytest>=7.4.0
pytest-asyncio>=0.23.5 